Uploading headers
Headers created:
· enemies.h :
           -Enemy_Create
           -Enemies_Draw
           -Enemies_Move
           -Enemies_Shooting
           -Enemies_DrawFires
           -Enemies_MovingFires
           -Enemies_HowManyFires
· sprites.h
           -sfSprite_createFromFile
           -sfSprite_CollisionPoint
           -outOfScreen
           -drawLifes
· structs.h: all the structs used in this code
