+++ megamania.cbp
@@ -46,12 +46,21 @@
 			<Add library="../CSFML/lib/gcc/libcsfml-system.a" />
 			<Add directory="../CSFML/lib/gcc" />
 		</Linker>
+		<Unit filename="enemies.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="enemies.h" />
 		<Unit filename="layout.c">
 			<Option compilerVar="CC" />
 		</Unit>
 		<Unit filename="main.c">
 			<Option compilerVar="CC" />
 		</Unit>
+		<Unit filename="sprites.c">
+			<Option compilerVar="CC" />
+		</Unit>
+		<Unit filename="sprites.h" />
+		<Unit filename="structs.h" />
 		<Extensions>
 			<envvars />
 			<code_completion />
